Add files via upload
Objective: In this project, students will employ data preprocessing techniques, build a decision tree classifier, and optimize the model using cross-validation and grid search. The aim is to understand the impact of various student attributes on their performance.
Dataset: The dataset student-por.csv consists of student attributes and their grades. It is available at the UCI ML Repository. You are only working with the Portuguese grades.